Complete tasks for 05-flex-modal.
I had to check the solution.css file for this because I was certain I was making things more complicated.  I removed all my initial changes and then worked through the solution step-by-step to make sure I understood why it was behaving that way.  I believe that I am beginning to understand that these solutions do not have to be super complex.